adds tests for standalone children, refators sib checking, optimizes unmounts, refactors subdiff/clean

+++ src/thread.js
@@ -1,4 +1,4 @@
-const { isArr } = require("./util")
+const { isArr, sib } = require("./util")
 
 /* a 'thread' is linked list of contiguous child chains.
 
@@ -88,33 +88,30 @@ const link = (e, f, p, s, next) => {
 
 const isAdd = f => (f = f && f.evt) && f.upd && !f.temp;
 
-const add = (f, p, s) => {
-  if (f.root < 2) isAdd(p) || queue(f, s, s ? s.sib : p && p.next);
-  else pushLeader(f);
-}
+const add = (f, p, s) =>
+  sib(f) ? isAdd(p) || queue(f, s, s ? s.sib : p && p.next) : pushLeader(f);
+
 const receive = (f, e=f.evt) => {
   if (!e.upd){
-    queue(f, f.prev, f.sib);
+    sib(f) ? queue(f, f.prev, f.sib) : pushLeader(f)
     e.temp = f.temp;
     e.upd = true
   }
 }
 const move = (f, p, s, ps, e=f.evt) => {
   isAdd(p) || dequeue(f, ps);
-  if (!e.upd){
-    e.temp = f.temp;
-    e.upd = true
-  }
+  if (!e.upd) e.temp = f.temp, e.upd = true;
   isAdd(p) || queue(f, s, s ? s.sib : p && p.next);
 }
 const remove = (f, p, e=f.evt) => {
   if (!e.upd || e.temp){
     e.temp = e.temp || f.temp;
     rems.push(f);
-    if (e.parent = f.root < 2 ? p : null) unlink(e, p.evt);
+    if (e.next = sib(f) && p) 
+      p.path > -2 && unlink(e, p.evt);
   }
-  if (f.root < 2) isAdd(p) || dequeue(f, f.prev);
-  else if (e.upd) popLeader(f);
+  sib(f) ? isAdd(p) || dequeue(f, f.prev) : e.upd && popLeader(f);
+  e.upd = false;
 }
 const emit = (eff, type, f, p, s, ps) => {
   if (isArr(eff)) for (eff of eff) eff[type] && eff[type](f, p, s, ps);
@@ -123,34 +120,32 @@ const emit = (eff, type, f, p, s, ps) => {
 // iterates and destroys the threads
 const flush = (c=0, f, e, p, owner) => {
   while(f = rems[c++])
-    emit((e = f.evt).effs, "willRemove", f, e.parent, e.prev, e.temp, f.evt = null);
+    emit((e = f.evt).effs, "willRemove", f, e.next, e.prev, e.temp, f.evt = null);
   rems.length = 0;
   if (!(f = head)) return;
   owner = f.parent;
   while(f) {
-    p = f.parent;
+    p = f.parent, c = sib(f);
     if ((e = f.evt).upd){
       e.upd = false;
       if (!e.temp){
-        emit(e.effs, "willAdd", f, f.root < 2 && p, f.root < 2 && f.prev, f.temp);
-        if (f.root < 2 && p) link(e, f, p.evt, f.prev);
-        if (p !== owner || f.next){
-          f = f.next || f.sib || p;
+        emit(e.effs, "willAdd", f, c && p, c && f.prev, f.temp);
+        if (c && p) link(e, f, p.evt, f.prev);
+        if (f.next){
+          f = f.next;
           continue;
         }
       } else {
         if (f.temp !== e.temp) emit(e.effs, "willReceive", f, f.temp);
-        if (f.prev !== e.prev){
+        if (c && f.prev !== e.prev){
           emit(e.effs, "willMove", f, p, e.prev, f.prev);
           unlink(e, p.evt), link(e, f, p.evt, f.prev);
         }
         e.temp = null;
       }
-    } else if (p !== owner) {
-      f = f.sib || p;
-      continue;
     }
-    if (f.root > 1 || !(f = f.sib) || !f.evt.upd){
+    if (p !== owner) f = f.sib || p;
+    else if (!c || !(f = f.sib) || !f.evt.upd){
       popLeader(head);
       if (f = head) owner = f.parent;
     }
